requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
